+++ src/alg/search/linear_search.h
@@ -1,6 +1,9 @@
 #ifndef LINEAR_H
 #define LINEAR_H
 
-int linear_search(int *arr, int size, int target);
+#include <stdlib.h>
+#include <sys/types.h>
+
+int linear_search(int *arr, ssize_t size, int target);
 
 #endif
