+++ src/alg/sort/bubble_sort.h
@@ -2,7 +2,8 @@
 #define BUBBLE_H
 
 #include <stdbool.h>
+#include <stdlib.h>
 
-bool bubble_sort(int *arr, int size);
+bool bubble_sort(int *arr, size_t size);
 
 #endif
