+++ src/alg/sort/merge_sort.h
@@ -2,9 +2,10 @@
 #define MERGE_H
 
 #include <stdbool.h>
+#include <stdlib.h>
 
 #include "../../utils/array_utils.h"
 
-bool merge_sort(int *arr, int start, int end);
+bool merge_sort(int *arr, size_t start, size_t end);
 
 #endif
