+++ src/alg/sort/selection_sort.h
@@ -2,7 +2,8 @@
 #define SELECTION_H
 
 #include <stdbool.h>
+#include <stdlib.h>
 
-bool selection_sort(int *arr, int size);
+bool selection_sort(int *arr, size_t size);
 
 #endif
